Split UriSearchExplanation from new trait SearchExplanation
