+++ kifi-backend/modules/search/app/com/keepit/search/engine/uri/UriSearchNonUserImpl.scala
@@ -5,12 +5,11 @@ import com.keepit.common.db.Id
 import com.keepit.common.logging.Logging
 import com.keepit.model._
 import com.keepit.search._
-import com.keepit.search.engine.explain.{ Explanation }
 import com.keepit.search.engine.result._
 import com.keepit.search.engine.{ ScoreContext, QueryEngine, QueryEngineBuilder, SearchTimeLogs }
 import com.keepit.search.index.Searcher
 
-import scala.concurrent.{ Future, Promise }
+import scala.concurrent.Future
 
 class UriSearchNonUserImpl(
     numHitsToReturn: Int,
@@ -84,7 +83,7 @@ class UriSearchNonUserImpl(
     }
   }
 
-  def explain(uriId: Id[NormalizedURI]): Explanation = {
+  def explain(uriId: Id[NormalizedURI]): UriSearchExplanation = {
     val engine = engineBuilder.build()
     val labels = engineBuilder.getQueryLabels()
     val query = engine.getQuery()
@@ -92,6 +91,6 @@ class UriSearchNonUserImpl(
 
     executeTextSearch(engine, collector)
 
-    Explanation(query, labels, collector.getMatchingValues(), collector.getBoostValues(), collector.getRawScore, collector.getBoostedScore, collector.getScoreComputation, collector.getDetails())
+    UriSearchExplanation(query, labels, collector.getMatchingValues(), collector.getBoostValues(), collector.getRawScore, collector.getBoostedScore, collector.getScoreComputation, collector.getDetails())
   }
 }
