+++ kifi-backend/modules/search/app/views/admin/explainResult.scala.html
@@ -1,4 +1,4 @@
-@(userId: Id[User], uriId: Id[NormalizedURI], explanation: Option[com.keepit.search.engine.explain.Explanation])
+@(userId: Id[User], uriId: Id[NormalizedURI], explanation: Option[com.keepit.search.engine.uri.UriSearchExplanation])
 
 <html>
