+++ kifi-backend/modules/search/test/com/keepit/search/controllers/FixedResultUriSearchCommander.scala
@@ -1,6 +1,6 @@
 package com.keepit.search.controllers
 
-import com.keepit.search.engine.uri.{ UriShardResult, UriSearchResult }
+import com.keepit.search.engine.uri.{ UriShardResult, UriSearchResult, UriSearchExplanation }
 import com.keepit.search.index._
 import com.keepit.common.util.Configuration
 import com.keepit.search._
@@ -11,7 +11,6 @@ import com.keepit.model._
 import scala.concurrent.Future
 import com.keepit.search.result.DecoratedResult
 import com.keepit.search.index.sharding.Shard
-import com.keepit.search.engine.explain.Explanation
 
 case class FixedResultIndexModule() extends IndexModule {
   var volatileDirMap = Map.empty[(String, Shard[_]), IndexDirectory] // just in case we need to reference a volatileDir. e.g. in spellIndexer
@@ -93,7 +92,7 @@ class FixedResultUriSearchCommander extends UriSearchCommander {
     predefinedConfig: Option[SearchConfig],
     debug: Option[String]): Future[UriShardResult] = ???
 
-  def explain(userId: Id[User], uriId: Id[NormalizedURI], lang: Option[String], experiments: Set[ExperimentType], query: String, debug: Option[String]): Future[Option[Explanation]] = ???
+  def explain(userId: Id[User], uriId: Id[NormalizedURI], lang: Option[String], experiments: Set[ExperimentType], query: String, debug: Option[String]): Future[Option[UriSearchExplanation]] = ???
   def warmUp(userId: Id[User]): Unit = {}
   def findShard(uriId: Id[NormalizedURI]): Option[Shard[NormalizedURI]] = ???
 }
